% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable.R
\name{lifetable}
\alias{lifetable}
\title{Construct a lifetable}
\usage{
lifetable(age, Dx, Px, age_group_length = c(1, 4, rep(5, ngroup - 2)))
}
\arguments{
\item{age}{A vector of ages or age ranges, e.g. '0-1', '1-4', '5-9', etc}

\item{Dx}{A numeric vector of death counts for the age groups.}

\item{Px}{A numeric vector of population sizes for the age groups.}

\item{age_group_length}{A numeric vector giving the length of each age group interval. The default values are for an infant-mortality value of 1, an early-childhood mortality of 4, and then 5 year age bins.}
}
\value{
A dataframe containing a completed life-table.
The column 'age' returns the age groups.
Column 'mx' is the mortality rate at age x.
Column 'ax' is the average number of person-years lived by individuals that die within the age range of x, e.g. individuals that die between '1-4' for age '1-4'.
Column 'qx' is the probability of dying at age x.
Column 'lx' is the magnitude of the population at x. It has a fixed radix of 1.
Column 'dx' is the number of lifetable deaths.
Column 'Lx' is the number of person years lived in age bin x.
Column 'Tx' is the number of person years lived above age x.
Column 'ex' is life-expectancy at age x.
}
\description{
Construct a lifetable
}
\examples{
# Age groups
Age <-  c("0", "1-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34",
"35-39", "40-44", "45-49", "50-54", "55-59", "60-64", "65-69",
"70-74", "75-79", "80-84", "85-89", "90-94", "95+")

# Death counts
Dx <- c(110934, 17693, 9168, 7378, 12192, 13356, 14223, 19210, 29136, 42961, 64310,
        90630, 116759, 153504, 196682, 223796, 220065, 185305, 120414, 50299, 13887)

# mid year population
Px <- c(4126560, 16195304, 18659141, 16815965, 13287434, 10803165, 10870165,
        11951709, 12508316, 11567216, 10528878, 9696502, 8595947, 7111897, 6186763,
        4661136, 2977347, 1518206, 648581, 170653, 44551)

lifetable(age = Age, Dx = Dx, Px = Px)
}
